Add profile update routes

+++ app/Http/Controllers/AuthController.php
@@ -29,6 +29,22 @@ class AuthController extends Controller {
             'token' => $user->createToken($request->device_name)->plainTextToken
         ]);
     }
+    public function changePassword(Request $request) {
+        $user = auth()->user();
+        $validator = Validator::make($request->all(), [
+            'current_password' => 'required|string',
+            'password' => ['required', 'confirmed', Password::min(8)->letters()->mixedCase()->numbers()->symbols()->uncompromised()]
+        ]);
+        if ($validator->fails())
+            return self::unsuccessfulResponse($validator->errors());
+        if (!Hash::check($request->current_password, $user->password))
+            return self::unsuccessfulResponse([
+                'current_password' => ['Invalid Password.']
+            ]);
+        $user->password = Hash::make($request->password);
+        $user->save();
+        return self::successfulResponse('Password Reset.');
+    }
     public function getToken(Request $request) {
         $validator = Validator::make($request->all(), [
             'email' => 'required|email|exists:users,email|min:5|max:255',
@@ -52,13 +68,24 @@ class AuthController extends Controller {
     public function getAllTokens(Request $request) {
         $user = auth()->user();
         return self::successfulResponse([
-            'tokens' => $user->tokens()
-                ->get()
-                ->map(fn($item) => ['device_name' => $item->name])
+            'tokens' => $user->tokens()->get()
+                ->map(fn($item) => [
+                    'device_name' => $item->name,
+                    'last_used_at' => $item->last_used_at
+                ])
         ]);
     }
     public function revokeAllTokens(Request $request) {
         $user = auth()->user();
+        $validator = Validator::make($request->all(), [
+            'password' => 'required|string'
+        ]);
+        if ($validator->fails())
+            return self::unsuccessfulResponse($validator->errors());
+        if (!Hash::check($request->password, $user->password))
+            return self::unsuccessfulResponse([
+                'password' => ['Invalid Password.']
+            ]);
         $user->tokens()->delete();
         return self::successfulResponse([
             'message' => 'All Tokens successfully deleted.'
